Relocate scdata classes
TODO: move TrgPlayer
