+++ eudplib/core/__init__.py
@@ -149,7 +149,6 @@ from .rawtrigger import (
     EncodeWeapon,
     Enemy,
     Exactly,
-    Flingy,
     Foes,
     Force1,
     Force2,
@@ -160,7 +159,6 @@ from .rawtrigger import (
     GiveUnits,
     HighestScore,
     Icon,
-    Image,
     Iscript,
     Kills,
     KillsAndRazings,
@@ -253,12 +251,10 @@ from .rawtrigger import (
     SetScore,
     SetSwitch,
     SetTo,
-    Sprite,
     StatText,
     Subtract,
     Switch,
     TalkingPortrait,
-    Tech,
     Toggle,
     Total,
     Transmission,
@@ -278,17 +274,13 @@ from .rawtrigger import (
     TrgSwitch,
     TrgSwitchAction,
     TrgSwitchState,
-    TrgUnit,
-    UnitOrder,
     Units,
     UnitsAndBuildings,
     UnMuteUnitSpeech,
     UnpauseGame,
     UnpauseTimer,
-    Upgrade,
     Victory,
     Wait,
-    Weapon,
 )
 from .variable import (
     EP_SetRValueStrictMode,
@@ -437,7 +429,6 @@ __all__ = [
     "EncodeWeapon",
     "Enemy",
     "Exactly",
-    "Flingy",
     "Foes",
     "Force1",
     "Force2",
@@ -448,7 +439,6 @@ __all__ = [
     "GiveUnits",
     "HighestScore",
     "Icon",
-    "Image",
     "Iscript",
     "Kills",
     "KillsAndRazings",
@@ -541,12 +531,10 @@ __all__ = [
     "SetScore",
     "SetSwitch",
     "SetTo",
-    "Sprite",
     "StatText",
     "Subtract",
     "Switch",
     "TalkingPortrait",
-    "Tech",
     "Toggle",
     "Total",
     "Transmission",
@@ -566,17 +554,13 @@ __all__ = [
     "TrgSwitch",
     "TrgSwitchAction",
     "TrgSwitchState",
-    "TrgUnit",
-    "UnitOrder",
     "Units",
     "UnitsAndBuildings",
     "UnMuteUnitSpeech",
     "UnpauseGame",
     "UnpauseTimer",
-    "Upgrade",
     "Victory",
     "Wait",
-    "Weapon",
     "EP_SetRValueStrictMode",
     "EUDCreateVariables",
     "EUDLightBool",
