+++ eudplib/core/rawtrigger/__init__.py
@@ -210,22 +210,14 @@ from .strenc import (
     EncodeUnitOrder,
     EncodeUpgrade,
     EncodeWeapon,
-    Flingy,
     Icon,
-    Image,
     Iscript,
     Portrait,
-    Sprite,
     StatText,
-    Tech,
     TrgAIScript,
     TrgLocation,
     TrgString,
     TrgSwitch,
-    TrgUnit,
-    UnitOrder,
-    Upgrade,
-    Weapon,
 )
 from .triggerscope import (
     NextTrigger,
@@ -435,22 +427,14 @@ __all__ = [
     "EncodeUnitOrder",
     "EncodeUpgrade",
     "EncodeWeapon",
-    "Flingy",
     "Icon",
-    "Image",
     "Iscript",
     "Portrait",
-    "Sprite",
     "StatText",
-    "Tech",
     "TrgAIScript",
     "TrgLocation",
     "TrgString",
     "TrgSwitch",
-    "TrgUnit",
-    "UnitOrder",
-    "Upgrade",
-    "Weapon",
     "NextTrigger",
     "PopTriggerScope",
     "PushTriggerScope",
