+++ eudplib/prelude/__init__.py
@@ -20,13 +20,13 @@ from ..core import (
     # EUDStruct
     EUDStruct, EUDVArray, selftype,
     # ConstType
-    Weapon, Flingy, Sprite, Image, Iscript,
-    Portrait, Icon, Upgrade, Tech,
-    UnitOrder, StatText,
+    Iscript,
+    Portrait, Icon,
+    StatText,
     TrgAIScript, TrgAllyStatus, TrgComparison, TrgCount,
     TrgLocation, TrgModifier, TrgOrder, TrgPlayer, TrgProperty,
     TrgPropState, TrgResource, TrgScore, TrgString, TrgSwitch,
-    TrgSwitchAction, TrgSwitchState, TrgUnit,
+    TrgSwitchAction, TrgSwitchState,
     # TrgPlayer
     P1, P2, P3, P4, P5, P6, P7, P8, P9, P10, P11, P12,
     Force1, Force2, Force3, Force4, AllPlayers, CurrentPlayer,
@@ -94,6 +94,17 @@ from ..core import (
     Clear, Random,
     # TrgCount
     All,
+)
+from ..offsetmap import (
+    CUnit,
+    Flingy,
+    Image,
+    Sprite,
+    Tech,
+    TrgUnit,
+    UnitOrder,
+    Upgrade,
+    Weapon,
 )
 
 __all__ = [
@@ -183,4 +194,5 @@ __all__ = [
     "Clear", "Random",
     # TrgCount
     "All",
+    "CUnit",
 ]
